requirements.txt
Flask~=3.1
Flask-SQLAlchemy~=3.1
python-dotenv~=1.1
